Conflicts:
	railo-java/railo-core/src/railo/runtime/Info.ini
	railo-java/railo-core/src/railo/runtime/PageContextImpl.java
	railo-java/railo-core/src/railo/runtime/type/UDFPlus.java
	railo-java/railo-core/src/resource/context/railo-context.ra
	railo-java/railo-core/src/resource/library/tag/Dump.cfc

+++ railo-java/railo-core/src/railo/runtime/converter/JSONConverter.java
@@ -503,7 +503,7 @@ public final class JSONConverter extends ConverterSupport {
 		// Number
 		if(object instanceof Number) {
 		    sb.append(goIn());
-		    sb.append(Caster.toString(((Number)object).doubleValue()));
+		    sb.append(Caster.toString(((Number)object)));
 		    return;
 		}
 		// Boolean
